feat: add route to delete a sauce

+++ src/controllers/sauces.controller.js
@@ -29,7 +29,7 @@ exports.createSauce = async (req, res, next) => {
   try {
     const sauce = new Sauce({
       ...JSON.parse(req.body.sauce),
-      imageUrl: `${req.protocol}://${req.get('host')}/${req.file.path}`
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     })
     await sauce.save()
 
@@ -51,17 +51,36 @@ exports.updateSauce = async (req, res, next) => {
       req.file
         ? {
             ...JSON.parse(req.body.sauce),
-            imageUrl: `${req.protocol}://${req.get('host')}/${req.file.path}`
+            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
           }
         : req.body,
       { runValidators: true }
     )
 
-    await fs.unlink(`public/images/${sauce.imageUrl.split('/').pop()}`)
+    if (req.file) {
+      await fs.unlink(`public/images/${sauce.imageUrl.split('/').pop()}`)
+    }
 
     res.status(200).send({ message: 'sauce updated successfully' })
   } catch (err) {
     if (req.file) await fs.unlink(req.file.path)
+    next(err)
+  }
+}
+
+exports.deleteSauce = async (req, res, next) => {
+  const { id } = req.params
+
+  try {
+    const sauce = await Sauce.findById(id)
+    if (!sauce) throw createError(404, 'sauce not found')
+
+    await sauce.deleteOne()
+
+    await fs.unlink(`public/images/${sauce.imageUrl.split('/').pop()}`)
+
+    res.status(200).json('sauce deleted successfully')
+  } catch (err) {
     next(err)
   }
 }
